Fix multiple wiki notes being passed in vimwiki cmd
The wiki notes were being passed twice as arguments to the vim command.
This was leading to multiple instances of file being opened. Now this
wouldn't have caused any race issues when writing, because both
instances where in the same vim session. But this fix ensures the path
is being passed only once.